Randomised sudoku maps, menu and retry buttons
Game loads the json file when window loads
When player chooses a level of difficulty a random
list of the 20 maps for that level is loaded into
an array. The first map from the randomed list is
displayed to the player.

A retry button and menu button are added to the
PlayGame scene.
If the player clicks on the menu button, then
game ends and the player is taken back to the
MenuGame scene.